Changing to trunk

+++ _____init.php
@@ -5,7 +5,7 @@ Contributors: 	Andrew Norcross (@norcross / andrewnorcross.com)
 				Jared Atchison (@jaredatch / jaredatchison.com)
 				Bill Erickson (@billerickson / billerickson.net)
 Description: 	This will create metaboxes with custom fields that will blow your mind.
-Version: 		0.6.1
+Version: 		0.5
 */
 
 /**
@@ -31,8 +31,6 @@ Version: 		0.6.1
 /************************************************************************
 		You should not edit the code below or things might explode!
 *************************************************************************/
-
-$meta_boxes = array();
 $meta_boxes = apply_filters ( 'cmb_meta_boxes' , $meta_boxes );
 foreach ( $meta_boxes as $meta_box ) {
 	$my_box = new cmb_Meta_Box( $meta_box );
@@ -87,9 +85,6 @@ class cmb_Meta_Box {
 
 		add_action( 'admin_menu', array(&$this, 'add') );
 		add_action( 'save_post', array(&$this, 'save') );
-
-		add_filter( 'cmb_show_on', array(&$this, 'add_for_id' ), 10, 2 );
-		add_filter( 'cmb_show_on', array(&$this, 'add_for_page_template' ), 10, 2 );
 	}
 
 	function add_post_enctype() {
@@ -106,82 +101,24 @@ class cmb_Meta_Box {
 	function add() {
 		$this->_meta_box['context'] = empty($this->_meta_box['context']) ? 'normal' : $this->_meta_box['context'];
 		$this->_meta_box['priority'] = empty($this->_meta_box['priority']) ? 'high' : $this->_meta_box['priority'];
-		$this->_meta_box['show_on'] = empty( $this->_meta_box['show_on'] ) ? array('key' => false, 'value' => false) : $this->_meta_box['show_on'];
-		
 		foreach ( $this->_meta_box['pages'] as $page ) {
-			if( apply_filters( 'cmb_show_on', true, $this->_meta_box ) )
-				add_meta_box( $this->_meta_box['id'], $this->_meta_box['title'], array(&$this, 'show'), $page, $this->_meta_box['context'], $this->_meta_box['priority']) ;
+			add_meta_box( $this->_meta_box['id'], $this->_meta_box['title'], array(&$this, 'show'), $page, $this->_meta_box['context'], $this->_meta_box['priority']) ;
 		}
 	}
-	
-	/**
-	 * Show On Filters
-	 * Use the 'cmb_show_on' filter to further refine the conditions under which a metabox is displayed.
-	 * Below you can limit it by ID and page template
-	 */
-	 
-	// Add for ID 
-	function add_for_id( $display, $meta_box ) {
-		if ( 'id' !== $meta_box['show_on']['key'] )
-			return $display;
 
-		// If we're showing it based on ID, get the current ID					
-		if( isset( $_GET['post'] ) ) $post_id = $_GET['post'];
-		elseif( isset( $_POST['post_ID'] ) ) $post_id = $_POST['post_ID'];
-		if( !isset( $post_id ) )
-			return $display;
-		
-		// If value isn't an array, turn it into one	
-		$meta_box['show_on']['value'] = !is_array( $meta_box['show_on']['value'] ) ? array( $meta_box['show_on']['value'] ) : $meta_box['show_on']['value'];
-		
-		// If current page id is in the included array, display the metabox
-
-		if ( in_array( $post_id, $meta_box['show_on']['value'] ) )
-			return true;
-		else
-			return false;
-	}
-	
-	// Add for Page Template
-	function add_for_page_template( $display, $meta_box ) {
-		if( 'page-template' !== $meta_box['show_on']['key'] )
-			return $display;
-			
-		// Get the current ID
-		if( isset( $_GET['post'] ) ) $post_id = $_GET['post'];
-		elseif( isset( $_POST['post_ID'] ) ) $post_id = $_POST['post_ID'];
-		if( !( isset( $post_id ) || is_page() ) ) return $display;
-			
-		// Get current template
-		$current_template = get_post_meta( $post_id, '_wp_page_template', true );
-		
-		// If value isn't an array, turn it into one	
-		$meta_box['show_on']['value'] = !is_array( $meta_box['show_on']['value'] ) ? array( $meta_box['show_on']['value'] ) : $meta_box['show_on']['value'];
-
-		// See if there's a match
-		if( in_array( $current_template, $meta_box['show_on']['value'] ) )
-			return true;
-		else
-			return false;
-	}
-	
 	// Show fields
 	function show() {
-	// $wp_version used for compatibility with new wp_editor() function
-		global $post, $wp_version;
+		global $post;
 
 		// Use nonce for verification
 		echo '<input type="hidden" name="wp_meta_box_nonce" value="', wp_create_nonce( basename(__FILE__) ), '" />';
 		echo '<table class="form-table cmb_metabox">';
 
 		foreach ( $this->_meta_box['fields'] as $field ) {
-			// Set up blank or default values for empty ones
-			if ( !isset( $field['name'] ) ) $field['name'] = '';
-			if ( !isset( $field['desc'] ) ) $field['desc'] = '';
-			if ( !isset( $field['std'] ) ) $field['std'] = '';
-			if ( 'file' == $field['type'] && !isset( $field['allow'] ) ) $field['allow'] = array( 'url', 'attachment' );
-			if ( 'file' == $field['type'] && !isset( $field['save'] ) )  $field['save']  = array( 'url' );
-						
+			// Set up blank values for empty ones
+			if ( !isset($field['desc']) ) $field['desc'] = '';
+			if ( !isset($field['std']) ) $field['std'] = '';
+			
 			$meta = get_post_meta( $post->ID, $field['id'], 'multicheck' != $field['type'] /* If multicheck this can be multiple values */ );
 
 			echo '<tr>';
@@ -194,9 +131,8 @@ class cmb_Meta_Box {
 				}			
 				echo '<td>';
 			}		
-						
+			
 			switch ( $field['type'] ) {
-
 				case 'text':
 					echo '<input type="text" name="', $field['id'], '" id="', $field['id'], '" value="', $meta ? $meta : $field['std'], '" style="width:97%" />','<p class="cmb_metabox_description">', $field['desc'], '</p>';
 					break;
@@ -230,7 +166,6 @@ class cmb_Meta_Box {
 					echo '<p class="cmb_metabox_description">', $field['desc'], '</p>';
 					break;
 				case 'radio_inline':
-					if( empty( $meta ) && !empty( $field['std'] ) ) $meta = $field['std'];
 					echo '<div class="cmb_radio_inline">';
 					foreach ($field['options'] as $option) {
 						echo '<div class="cmb_radio_inline_option"><input type="radio" name="', $field['id'], '" value="', $option['value'], '"', $meta == $option['value'] ? ' checked="checked"' : '', ' />', $option['name'], '</div>';
@@ -239,7 +174,6 @@ class cmb_Meta_Box {
 					echo '<p class="cmb_metabox_description">', $field['desc'], '</p>';
 					break;
 				case 'radio':
-					if( empty( $meta ) && !empty( $field['std'] ) ) $meta = $field['std'];
 					foreach ($field['options'] as $option) {
 						echo '<p><input type="radio" name="', $field['id'], '" value="', $option['value'], '"', $meta == $option['value'] ? ' checked="checked"' : '', ' />', $option['name'].'</p>';
 					}
@@ -264,16 +198,10 @@ class cmb_Meta_Box {
 					echo '<p class="cmb_metabox_description">', $field['desc'], '</p>';
 					break;
 				case 'wysiwyg':
-					/* Make sure that the new wp_editor() function is available.
-					 * Otherwise, use the "old" version of the WYSIWYG editor */
-					if( function_exists( 'wp_editor' ) ) {
-						wp_editor( $meta ? $meta : $field['std'], $field['id'] );
-					} else {
-						echo '<div id="poststuff" class="meta_mce">';
-						echo '<div class="customEditor"><textarea name="', $field['id'], '" id="', $field['id'], '" cols="60" rows="7" style="width:97%">', $meta ? wpautop($meta, true) : '', '</textarea></div>';
-						echo '</div>';
-					}
-			        	echo '<p class="cmb_metabox_description">', $field['desc'], '</p>';
+					echo '<div id="poststuff" class="meta_mce">';
+					echo '<div class="customEditor"><textarea name="', $field['id'], '" id="', $field['id'], '" cols="60" rows="7" style="width:97%">', $meta ? wpautop($meta, true) : '', '</textarea></div>';
+					echo '</div>';
+			        echo '<p class="cmb_metabox_description">', $field['desc'], '</p>';
 					break;
 				case 'taxonomy_select':
 					echo '<select name="', $field['id'], '" id="', $field['id'], '">';
@@ -301,18 +229,6 @@ class cmb_Meta_Box {
 					}
 					echo '<p class="cmb_metabox_description">', $field['desc'], '</p>';
 					break;
-				case 'taxonomy_multicheck':
-					echo '<ul>';
-					$names = wp_get_object_terms( $post->ID, $field['taxonomy'] );
-					$terms = get_terms( $field['taxonomy'], 'hide_empty=0' );
-					foreach ($terms as $term) {
-						echo '<li><input type="checkbox" name="', $field['id'], '[]" id="', $field['id'], '" value="', $term->name , '"'; 
-						foreach ($names as $name) {
-							if ( $term->slug == $name->slug ){ echo ' checked="checked" ';};
-						}
-						echo' /><label>', $term->name , '</label></li>';
-					}
-				break;
 				case 'file_list':
 					echo '<input id="upload_file" type="text" size="36" name="', $field['id'], '" value="" />';
 					echo '<input class="upload_button button" type="button" value="Upload File" />';
@@ -336,12 +252,8 @@ class cmb_Meta_Box {
 							}
 						break;
 				case 'file':
-					$input_type_url = "hidden";
-					if ( 'url' == $field['allow'] || ( is_array( $field['allow'] ) && in_array( 'url', $field['allow'] ) ) )
-						$input_type_url="text";
-					echo '<input class="upload_file" type="' . $input_type_url . '" size="45" id="', $field['id'], '" name="', $field['id'], '" value="', $meta, '" />';
+					echo '<input id="upload_file" type="text" size="45" class="', $field['id'], '" name="', $field['id'], '" value="', $meta, '" />';
 					echo '<input class="upload_button button" type="button" value="Upload File" />';
-					echo '<input class="upload_file_id" type="hidden" id="', $field['id'], '_id" name="', $field['id'], '_id" value="', get_post_meta( $post->ID, $field['id'] . "_id",true), '" />';					
 					echo '<p class="cmb_metabox_description">', $field['desc'], '</p>';
 					echo '<div id="', $field['id'], '_status" class="cmb_upload_status">';	
 						if ( $meta != '' ) { 
@@ -363,8 +275,8 @@ class cmb_Meta_Box {
 				break;
 				default:
 					do_action('cmb_render_' . $field['type'] , $field, $meta);
+
 			}
-			
 			echo '</td>','</tr>';
 		}
 		echo '</table>';
@@ -372,7 +284,6 @@ class cmb_Meta_Box {
 
 	// Save data from metabox
 	function save( $post_id)  {
-		global $wp_version;
 		// verify nonce
 		if ( ! isset( $_POST['wp_meta_box_nonce'] ) || !wp_verify_nonce( $_POST['wp_meta_box_nonce'], basename(__FILE__) ) ) {
 			return $post_id;
@@ -397,15 +308,18 @@ class cmb_Meta_Box {
 			$old = get_post_meta( $post_id, $name, 'multicheck' != $field['type'] /* If multicheck this can be multiple values */ );
 			$new = isset( $_POST[$field['id']] ) ? $_POST[$field['id']] : null;
 
-			// wpautop() should not be needed with version 3.3 and later
-			if ( $field['type'] == 'wysiwyg' && !function_exists( 'wp_editor' ) ) {
+			if ( $field['type'] == 'wysiwyg' ) {
 				$new = wpautop($new);
 			}
 			
-			if ( in_array( $field['type'], array( 'taxonomy_select', 'taxonomy_radio', 'taxonomy_multicheck' ) ) )  {	
+			if ( $field['type'] == 'taxonomy_select' )  {	
 				$new = wp_set_object_terms( $post_id, $new, $field['taxonomy'] );	
 			}
 			
+			if ( $field['type'] == 'taxonomy_radio' )  {		
+				$new = wp_set_object_terms( $post_id, $new, $field['taxonomy'] );
+			}
+
 			if ( ($field['type'] == 'textarea') || ($field['type'] == 'textarea_small') ) {
 				$new = htmlspecialchars( $new );
 			}
@@ -414,8 +328,8 @@ class cmb_Meta_Box {
 				$new = strtotime( $new );
 			}
 			
-			$new = apply_filters('cmb_validate_' . $field['type'], $new, $post_id, $field);			
-			
+			$new = apply_filters('cmb_validate_' . $field['type'], $new, $post_id, $field);
+
 			// validate meta value
 			if ( isset( $field['validate_func']) ) {
 				$ok = call_user_func( array( 'cmb_Meta_Box_Validate', $field['validate_func']), $new );
@@ -438,25 +352,9 @@ class cmb_Meta_Box {
 				}
 			} elseif ( $new && $new != $old ) {
 				update_post_meta( $post_id, $name, $new );
-			} elseif ( '' == $new && $old ) {
+			} elseif ( '' == $new && $old && $field['type'] != 'file' ) {
 				delete_post_meta( $post_id, $name, $old );
 			}
-			
-			if ( 'file' == $field['type'] ) {
-				$name = $field['id'] . "_id";
-				$old = get_post_meta( $post_id, $name, 'multicheck' != $field['type'] /* If multicheck this can be multiple values */ );
-				if ( isset( $field['save'] ) && is_array($field['save']) && in_array('id',$field['save'] )) {
-					$new = isset( $_POST[$name] ) ? $_POST[$name] : null;
-				} else {
-					$new = "";
-				}
-
-				if ( $new && $new != $old ) {
-					update_post_meta( $post_id, $name, $new );
-				} elseif ( '' == $new && $old ) {
-					delete_post_meta( $post_id, $name, $old );
-				}
-			}			
 		}
 	}
 }
@@ -511,16 +409,7 @@ function cmb_editor_footer_scripts() { ?>
  			});
 		});
 	/* ]]> */</script>
-	<?php if ( isset( $_GET['cmb_force_send'] ) && 'true' == $_GET['cmb_force_send'] ) { 
-		$label = $_GET['cmb_send_label']; 
-		if ( empty( $label ) ) $label="Select File";?>	
-	<script type="text/javascript">
-		jQuery(function($) {
-			$('td.savesend input').val('<?php echo $label; ?>');
-		});
-	</script>
-	<?php } ?>	
-<?php }
+	<?php }
 add_action( 'admin_print_footer_scripts', 'cmb_editor_footer_scripts', 99 );
 
 function cmb_styles_inline() { 
@@ -538,7 +427,7 @@ function cmb_styles_inline() {
 		table.cmb_metabox input, table.cmb_metabox textarea { font-size:11px; padding: 5px;}
 		table.cmb_metabox li { font-size:11px; }
 		table.cmb_metabox ul { padding-top:5px; }
-		table.cmb_metabox select { font-size:11px;}
+		table.cmb_metabox select { font-size:11px; padding: 5px 10px;}
 		table.cmb_metabox input:focus, table.cmb_metabox textarea:focus { background: #fffff8;}
 		.cmb_metabox_title { margin: 0 0 5px 0; padding: 5px 0 0 0; font: italic 24px/35px Georgia,"Times New Roman","Bitstream Charter",Times,serif;}
 		.cmb_radio_inline { padding: 4px 0 0 0;}
@@ -557,57 +446,4 @@ function cmb_styles_inline() {
 	<?php
 }
 
-// Force 'Insert into Post' button from Media Library 
-add_filter( 'get_media_item_args', 'cmb_force_send' );
-function cmb_force_send( $args ) {
-		
-	// if the Gallery tab is opened from a custom meta box field, add Insert Into Post button	
-	if ( isset( $_GET['cmb_force_send'] ) && 'true' == $_GET['cmb_force_send'] )
-		$args['send'] = true;
-	
-	// if the From Computer tab is opened AT ALL, add Insert Into Post button after an image is uploaded	
-	if ( isset( $_POST['attachment_id'] ) && '' != $_POST["attachment_id"] ) {
-		
-		$args['send'] = true;		
-
-		// TO DO: Are there any conditions in which we don't want the Insert Into Post 
-		// button added? For example, if a post type supports thumbnails, does not support
-		// the editor, and does not have any cmb file inputs? If so, here's the first
-		// bits of code needed to check all that.
-		// $attachment_ancestors = get_post_ancestors( $_POST["attachment_id"] );
-		// $attachment_parent_post_type = get_post_type( $attachment_ancestors[0] );
-		// $post_type_object = get_post_type_object( $attachment_parent_post_type );
-
-	}		
-	
-	// change the label of the button on the From Computer tab
-	if ( isset( $_POST['attachment_id'] ) && '' != $_POST["attachment_id"] ) {
-
-		echo '
-			<script type="text/javascript">
-				function cmbGetParameterByNameInline(name) {
-					name = name.replace(/[\[]/, "\\\[").replace(/[\]]/, "\\\]");
-					var regexS = "[\\?&]" + name + "=([^&#]*)";
-					var regex = new RegExp(regexS);
-					var results = regex.exec(window.location.href);
-					if(results == null)
-						return "";
-					else
-						return decodeURIComponent(results[1].replace(/\+/g, " "));
-				}
-							
-				jQuery(function($) {
-					if (cmbGetParameterByNameInline("cmb_force_send")=="true") {
-						var cmb_send_label = cmbGetParameterByNameInline("cmb_send_label");
-						$("td.savesend input").val(cmb_send_label);
-					}
-				});
-			</script>
-		';
-	}
-	 
-    return $args;
-
-}
-
-// End. That's it, folks! / 
+// End. That's it, folks! //
